requirementsPC.txt
aiohappyeyeballs
aiohttp
aiosignal
altair
annotated-types
anyio
AnyQt
anywidget
appdirs
arrow
astor
asttokens
attrs
backoff
baycomp
beautifulsoup4
bertopic
blinker
blis
bokeh
Bottleneck
cachetools
catalogue
catboost
cattrs
certifi~=2024.8.30
cffi
chardet
charset-normalizer
click
cloudpathlib
cloudpickle
comm
commonmark
confection
contourpy
cryptography
cycler
cymem
dask~=2024.11.1
dask-expr
dateutils
db-dtypes
debugpy
decorator
dictdiffer
distributed~=2024.11.1
distro
docutils
duckdb
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et-xmlfile
executing
fastjsonschema
filelock
fonttools
frozenlist
fsspec
gcsfs
gitdb
GitPython
google
google-api-core
google-auth
google-auth-oauthlib
google-cloud
google-cloud-bigquery
google-cloud-core
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
graphviz
greenlet
grpcio
grpcio-status
gw_dsl_parser
h11
hdbscan
httpcore
httpx
huggingface-hub
idna
importlib_resources
ipykernel
ipylab
ipython
ipywidgets
jaraco.classes
jaraco.context
jaraco.functools
jedi
jeepney
Jinja2
jiter
joblib
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
jupyterlab_widgets
kanaries_track
keyring
keyrings.alt
kiwisolver
langcodes
language_data
llvmlite
locket
lz4
marisa-trie
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
monotonic
more-itertools
mpmath
msgpack
multidict
murmurhash
mysql-connector
mysql-connector-python
narwhals
nbformat
nest-asyncio
networkx
nltk
numba
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
oauthlib
openai
openpyxl
openTSNE==1.0.1
orange-canvas-core==0.2.1
orange-widget-base==4.24.0
Orange3==3.37.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.2.2
plotly==5.22.0
polars==1.10.0
preshed==3.0.9
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.1.0
psycopg==3.1.19
psycopg2-binary==2.9.9
psygnal==0.11.1
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pygwalker==0.4.9.13
PyMySQL==1.1.1
pynndescent==0.5.13
pyparsing==3.2.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqtgraph==0.13.7
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyspark==3.5.3
python-dateutil==2.9.0.post0
python-louvain==0.16
pytz==2024.2
PyYAML==6.0.1
pyzmq==26.0.3
qasync==0.27.1
qtconsole==5.5.2
QtPy==2.4.1
quickjs==1.19.4
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-cache==1.2.0
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.21.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.4.2
scipy==1.13.1
seaborn==0.13.2
SecretStorage==3.3.3
segment-analytics-python==2.2.3
sentence-transformers==3.3.0
serverfiles==0.3.1
setuptools==75.3.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spark-nlp==5.5.1
sql==2022.4.0
SQLAlchemy==2.0.36
sqlglot==25.29.0
srsly==2.4.8
stack-data==0.6.3
streamlit==1.40.1
sympy==1.13.1
tblib==3.0.0
tenacity==8.3.0
thinc==8.3.2
threadpoolctl==3.5.0
tokenizers==0.20.3
toml==0.10.2
toolz==1.0.0
torch==2.5.1
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.2
triton==3.1.0
typer==0.13.0
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
umap-learn==0.5.7
url-normalize==1.4.3
urllib3==2.2.3
wasabi==1.1.3
wasmtime==25.0.0
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
wheel==0.43.0
widgetsnbextension==4.0.13
wordcloud==1.9.4
wrapt==1.16.0
xgboost==2.0.3
xlrd==2.0.1
XlsxWriter==3.2.0
xyzservices==2024.9.0
yarl==1.16.0
zict==3.0.0
